pyproject.toml
[build-system]
requires = ["hatchling>=1.21"]
build-backend = "hatchling.build"

[project]
name = "nimbus"
version = "0.1.0"
description = "Nimbus: control plane and host agent prototype for accelerated GitHub Actions"
authors = [
  { name = "EvalOps" }
]
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.112,<0.115",
  "uvicorn[standard]>=0.30,<0.32",
  "httpx>=0.27,<0.28",
  "pydantic>=2.7,<2.9",
  "pydantic-settings>=2.2,<2.3",
  "redis>=5.0,<5.1",
  "sqlalchemy>=2.0,<2.1",
  "asyncpg>=0.29,<0.30",
  "aiosqlite>=0.19,<0.20",
  "python-dotenv>=1.0,<1.1",
  "PyJWT[crypto]>=2.8,<2.9",
  "boto3>=1.35,<1.36",
  "structlog>=24.1,<26",
  "opentelemetry-api>=1.37,<1.38",
  "opentelemetry-sdk>=1.37,<1.38",
  "opentelemetry-exporter-otlp>=1.37,<1.38",
  "opentelemetry-instrumentation-fastapi>=0.58b0,<0.59"
]

[project.optional-dependencies]
host = [
  "structlog>=24.1,<26",
  "aiodocker>=0.21,<0.22"
]

[tool.hatch.build.targets.wheel]
packages = ["src/nimbus"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
